backend/requirements.txt
fastapi[all]>=0.115.12
langchain
langchain-google-genai
psycopg2-binary
python-dotenv
sqlalchemy
uvicorn